refactor: poweruser 로직 point 디렉토리에서 분리
- poweruser 로직 전체를 Poweruser 디렉토리 생성후 이동
- 기존에 poweruser를 사용했던 코드들의 경로 변경
