+++ src/main/java/com/chungnamthon/cheonon/home/dto/HomeResponse.java
@@ -2,7 +2,7 @@ package com.chungnamthon.cheonon.home.dto;
 
 import com.chungnamthon.cheonon.meeting.dto.response.MeetingPreviewResponse;
 import com.chungnamthon.cheonon.map.dto.AffiliateHomePreviewResponse;
-import com.chungnamthon.cheonon.point.dto.response.PowerUserResponse;
+import com.chungnamthon.cheonon.poweruser.dto.PowerUserResponse;
 import lombok.Builder;
 import lombok.Getter;
