+++ src/main/java/com/chungnamthon/cheonon/home/service/HomeService.java
@@ -5,8 +5,8 @@ import com.chungnamthon.cheonon.map.dto.AffiliateHomePreviewResponse;
 import com.chungnamthon.cheonon.map.service.AffiliateService;
 import com.chungnamthon.cheonon.meeting.dto.response.MeetingPreviewResponse;
 import com.chungnamthon.cheonon.meeting.repository.MeetingRepository;
-import com.chungnamthon.cheonon.point.dto.response.PowerUserResponse;
-import com.chungnamthon.cheonon.point.service.PowerUserService;
+import com.chungnamthon.cheonon.poweruser.dto.PowerUserResponse;
+import com.chungnamthon.cheonon.poweruser.service.PowerUserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
