+++ src/main/java/com/chungnamthon/cheonon/point/repository/PowerUserRepository.java
@@ -1,7 +1,6 @@
 package com.chungnamthon.cheonon.point.repository;
 
-import com.chungnamthon.cheonon.point.domain.PowerUser;
-import lombok.Builder;
+import com.chungnamthon.cheonon.poweruser.domain.PowerUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
