+++ src/main/java/com/chungnamthon/cheonon/poweruser/service/PowerUserService.java
@@ -1,7 +1,7 @@
-package com.chungnamthon.cheonon.point.service;
+package com.chungnamthon.cheonon.poweruser.service;
 
-import com.chungnamthon.cheonon.point.domain.PowerUser;
-import com.chungnamthon.cheonon.point.dto.response.PowerUserResponse;
+import com.chungnamthon.cheonon.poweruser.domain.PowerUser;
+import com.chungnamthon.cheonon.poweruser.dto.PowerUserResponse;
 import com.chungnamthon.cheonon.point.repository.PointRepository;
 import com.chungnamthon.cheonon.point.repository.PowerUserRepository;
 import com.chungnamthon.cheonon.user.domain.User;
